ZOOKEEPER-1182. Make findbugs usable in Eclipse (Thomas Koch via phunt)
git-svn-id: https://svn.apache.org/repos/asf/zookeeper/trunk@1171869 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/test/org/apache/zookeeper/test/ClientHammerTest.java
@@ -214,8 +214,8 @@ public class ClientHammerTest extends ClientBase {
         for (HammerThread h : threads) {
             final int safetyFactor = 3;
             verifyThreadTerminated(h,
-                    threads.length * childCount
-                    * HAMMERTHREAD_LATENCY * safetyFactor);
+                    (long)threads.length * (long)childCount
+                    * HAMMERTHREAD_LATENCY * (long)safetyFactor);
         }
         LOG.info(new Date() + " Total time "
                 + (System.currentTimeMillis() - start));
